refactor: split vec-related method
