+++ enum-table/src/impls/mod.rs
@@ -1,4 +1,7 @@
 mod std;
 
+mod vec;
+pub use vec::*;
+
 #[cfg(feature = "serde")]
 mod serde;
